+++ aten/src/ATen/PadNd.h
@@ -11,12 +11,16 @@ enum class padding_mode {
 
 static inline c10::string_view padding_mode_string(padding_mode m) {
   switch (m) {
-    case padding_mode::reflect: return "reflect";
-    case padding_mode::replicate: return "replicate";
-    case padding_mode::circular: return "circular";
-    case padding_mode::constant: return "constant";
+    case padding_mode::reflect:
+      return "reflect";
+    case padding_mode::replicate:
+      return "replicate";
+    case padding_mode::circular:
+      return "circular";
+    case padding_mode::constant:
+      return "constant";
   }
   TORCH_CHECK(false, "Invalid padding mode (", static_cast<int64_t>(m), ")");
 }
 
-}  // namespace at
+} // namespace at
